Move functions to image_utils.py
